more test coverage
