+++ serial_usbipclient/tests/common_test_base.py
@@ -9,9 +9,12 @@ from os import getenv
 from typing import Optional
 from unittest import TestCase
 
+from socket import AddressFamily, SocketKind
 from mock_usbip import MockUSBIP
 
-from serial_usbipclient.usbip_client import USBIPClient
+from serial_usbipclient.usbip_client import USBIPClient, SocketWrapper
+from serial_usbipclient.protocol.packets import BasicCommands, CommonHeader
+
 
 LOG_FORMAT: str = '%(asctime)s\t%(levelname)s \t[%(filename)s:%(lineno)d] - %(message)s'
 logging.basicConfig(
@@ -76,3 +79,51 @@ class CommonTestBase(TestCase):
             self.mock_usbip = None
 
         super().tearDown()
+
+
+class MockSocketWrapper(SocketWrapper):
+    """for injecting and managing a fake connection"""
+    def __init__(self, family: AddressFamily, kind: SocketKind):
+        """set up local variables"""
+        super().__init__(family, kind)
+        self._protocol_responses: dict[str, bytes] = {}
+        data_path: str = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'usbip_packets.json')
+        with open(file=data_path, mode='r', encoding='utf-8') as recording:
+            self._protocol_responses = json.loads(recording.read())
+        self.send_data: list[bytes] = []  # track commands that were sent
+        self.response_data: bytes = bytes()  # buffered responses based on commands sent
+
+    def connect(self, address: tuple[str, int]):
+        """mock connection"""
+        self._address = address
+
+    def getsockname(self) -> tuple[str, int]:
+        """return the socket name"""
+        return self._address
+
+    def shutdown(self, how: int) -> None:
+        """perform mock shutdown"""
+        return
+
+    def sendall(self, data: bytes) -> None:
+        """sending data to the device"""
+        self.send_data.append(data)
+        header: CommonHeader = CommonHeader.unpack(data)
+        response_key: str = ''
+        if header.command == BasicCommands.REQ_DEVLIST:  # asking for the device list
+            response_key = 'OP_REP_DEVLIST'
+        elif header.command == BasicCommands.REQ_IMPORT:
+            response_key = 'OP_REP_IMPORT'
+        if response_key:
+            self.response_data += bytes.fromhex("".join([item for item in self._protocol_responses[response_key]]))
+
+    def recv(self, size: int) -> bytes:
+        """return data corresponding to last send"""
+        if self.response_data:
+            if len(self.response_data) < size:
+                size = len(self.response_data)
+            response: bytes = self.response_data[:size]
+            self.response_data = self.response_data[size:]
+            return response
+
+        return bytes()
