+++ serial_usbipclient/tests/test_exceptions.py
@@ -1,7 +1,17 @@
 """test exceptions are properly raised"""
+import socket
 
-from common_test_base import CommonTestBase
-from serial_usbipclient.usbip_client import USBIPValueError, USBIP_Connection
+from common_test_base import CommonTestBase, MockSocketWrapper
+from serial_usbipclient.protocol.packets import CMD_SUBMIT, CMD_UNLINK
+from serial_usbipclient.protocol.urb_packets import ConfigurationDescriptor
+from serial_usbipclient.usbip_client import USBIPValueError, USBIP_Connection, USBConnectionLostError
+
+
+class ConnectionErrorSocketWrapper(MockSocketWrapper):
+    """raise a gai error on connection"""
+    def sendall(self, data: bytes) -> None:
+        """raise a socket error"""
+        raise ConnectionError('mock socket wrapper')
 
 
 class TestExceptions(CommonTestBase):
@@ -20,3 +30,16 @@ class TestExceptions(CommonTestBase):
 
         with self.assertRaisesRegex(expected_exception=USBIPValueError, expected_regex='no endpoint!'):
             _ = conn.pending_reads
+
+        with self.assertRaisesRegex(expected_exception=USBIPValueError, expected_regex='no configuration'):
+            _ = conn.configuration
+
+        with self.assertRaisesRegex(expected_exception=USBIPValueError, expected_regex='cannot set configuration'):
+            conn.configuration = ConfigurationDescriptor()
+
+        with self.assertRaisesRegex(expected_exception=USBIPValueError, expected_regex='socket not available'):
+            conn.send_command(command=CMD_SUBMIT())
+
+        with self.assertRaisesRegex(expected_exception=USBConnectionLostError, expected_regex='connection lost'):
+            conn.socket = ConnectionErrorSocketWrapper(family=socket.AF_INET, kind=socket.SOCK_STREAM)
+            conn.send_unlink(command=CMD_UNLINK(seqnum=2, unlink_seqnum=1))
